Improve CNAME file generation
Only include a CNAME file in the output build, when the user is
using a custom domain, as configured in src/data/meta.json.
This conditional inclusion of files requires the use of JavaScript
templates (and a few hacks).

11ty allows to exclude files by setting `permalink` to false in the
document frontmatter. 11ty can also accept permalink functions
(https://www.11ty.dev/docs/languages/javascript/#permalink-function)
that generate dynamic permalinks. However, returning false from
these does not seem to work, so instead the code introduced in this
change just reads the configuration directly from the source file.
It then decides, whether or not to include a CNAME file with the
specified custom domain name.